Algorithm.java, Item.java
Implemented the algorithm for calculating the bill and applying the tax and tip to each item in the bill.

Added base_amount and total_amount to Item.java class fields to replace amount field because then we can calculate tax and tip and store base_amounts and total_amounts in the items themselves.